Use emulator_http_transport mod from ledger-testing crate
instead of adding one to the signer crate
