+++ cmd/crates/ledger-testing/Cargo.toml
@@ -6,3 +6,8 @@ version = "0.1.0"
 [dependencies]
 testcontainers = { git = "https://github.com/testcontainers/testcontainers-rs.git", rev = "4b3e4f08a2c0bdf521636b03f959f004e6d216aa" }
 phf = { version = "0.11.2", features = ["macros"] }
+thiserror = "1.0.32"
+ledger-transport = "0.10.0"
+serde = "1.0.82"
+reqwest = { version = "0.12.4", features = ["json"]}
+hex = "0.4.3"
