+++ cmd/crates/ledger-testing/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod speculos;
+pub mod emulator_http_transport;
